Adds helper function for retrieval of input documents
I wrote this function and intended to extract inputs, call this
function, and then enrich the original inputs with the results. I then
realized that most of this could be done for us by specifying `_source:
true` in the riskiest_inputs subaggregation.

For now, this function is unused, but if in the future we need e.g.
runtime field support for these docs, we might need to reevaluate/use
this.