Add order-dependent integration test
I realized in the course of changing implementation that, because all of
the existing tests have the _same_ risk score value for all alerts, the
ordering/aggregation of them doesn't really matter.

This test does not have that robustness, and so if it fails, it means we
changed our sorting/summing logic.